Check style and fix project definitions and remove motor setpoints
