+++ base/uk.ac.stfc.isis.ibex.motor/src/uk/ac/stfc/isis/ibex/motor/internal/MotorsTable.java
@@ -63,6 +63,10 @@ public class MotorsTable extends Closer {
 		}
 	}
 	
+	/**
+	 * 
+	 * @return motors in this table
+	 */
 	public Collection<Motor> motors() {
 		return new ArrayList<>(motors);
 	}
