+++ base/uk.ac.stfc.isis.ibex.scriptgenerator/src/uk/ac/stfc/isis/ibex/scriptgenerator/generation/GeneratedLanguage.java
@@ -16,6 +16,10 @@ public enum GeneratedLanguage {
 	
 	private final String name;
 	
+	/**
+	 * Constructor
+	 * @param name name of the language
+	 */
 	GeneratedLanguage(String name) {
 		this.name = name;
 	}
