+++ base/uk.ac.stfc.isis.ibex.scriptgenerator/src/uk/ac/stfc/isis/ibex/scriptgenerator/generation/UnsupportedLanguageException.java
@@ -1,8 +1,16 @@
 package uk.ac.stfc.isis.ibex.scriptgenerator.generation;
 
+/**
+ * Exception thrown when the language requested is not supported by the script generator.
+ * 
+ */
 @SuppressWarnings("serial")
 public class UnsupportedLanguageException extends Exception {
 	
+    /**
+     * Constructor.
+     * @param errorMessage error message for log 
+     */
 	public UnsupportedLanguageException(String errorMessage) {
 		super(errorMessage);
 	}
