+++ base/uk.ac.stfc.isis.ibex.ui.motor.tests/src/uk/ac/stfc/isis/ibex/ui/motor/tests/TestMotor.java
@@ -24,14 +24,14 @@ package uk.ac.stfc.isis.ibex.ui.motor.tests;
 import uk.ac.stfc.isis.ibex.motor.Motor;
 import uk.ac.stfc.isis.ibex.motor.MotorDirection;
 import uk.ac.stfc.isis.ibex.motor.MotorEnable;
-import uk.ac.stfc.isis.ibex.motor.MotorSetpoint;
 
 public class TestMotor extends Motor {
 
     public String name = "Motor name";
     public String address = "Address";
     public String description = "Description";
-    public TestMotorSetpoint testMotorSetpoint = new TestMotorSetpoint();
+    public Double value = -1.23456;
+    public Double setpoint = 1.23456;
     public MotorEnable enabled = MotorEnable.ENABLE;
     public Double lowerLimit = -5.0;
     public Double upperLimit = 5.0;
@@ -62,8 +62,13 @@ public class TestMotor extends Motor {
     }
 
     @Override
-    public MotorSetpoint getSetpoint() {
-        return testMotorSetpoint;
+    public Double getValue() {
+        return value;
+    }
+
+    @Override
+    public Double getSetpoint() {
+        return setpoint;
     }
 
     @Override
