--- base/uk.ac.stfc.isis.ibex.ui.motor.tests/src/uk/ac/stfc/isis/ibex/ui/motor/tests/TestMotorSetpoint.java
@@ -1,58 +1,0 @@
- /*
- * This file is part of the ISIS IBEX application.
- * Copyright (C) 2012-2016 Science & Technology Facilities Council.
- * All rights reserved.
- *
- * This program is distributed in the hope that it will be useful.
- * This program and the accompanying materials are made available under the
- * terms of the Eclipse Public License v1.0 which accompanies this distribution.
- * EXCEPT AS EXPRESSLY SET FORTH IN THE ECLIPSE PUBLIC LICENSE V1.0, THE PROGRAM 
- * AND ACCOMPANYING MATERIALS ARE PROVIDED ON AN "AS IS" BASIS, WITHOUT WARRANTIES 
- * OR CONDITIONS OF ANY KIND.  See the Eclipse Public License v1.0 for more details.
- *
- * You should have received a copy of the Eclipse Public License v1.0
- * along with this program; if not, you can obtain a copy from
- * https://www.eclipse.org/org/documents/epl-v10.php or 
- * http://opensource.org/licenses/eclipse-1.0.php
- */
-
-package uk.ac.stfc.isis.ibex.ui.motor.tests;
-
-import uk.ac.stfc.isis.ibex.motor.MotorSetpoint;
-
-public class TestMotorSetpoint extends MotorSetpoint {
-
-    public Double value = -1.23456;
-    public Double setpoint = 1.23456;
-    public boolean canHome = false;
-
-    public TestMotorSetpoint() {
-        super("Motor name");
-    }
-
-    @Override
-    public Double getValue() {
-        return value;
-    }
-
-    @Override
-    public Double getSetpoint() {
-        return setpoint;
-    }
-
-    @Override
-    public void setSetpoint(double value) {
-        return;
-    }
-
-    @Override
-    public void home() {
-        return;
-    }
-
-    @Override
-    public Boolean canHome() {
-        return canHome;
-    }
-
-}
