+++ base/uk.ac.stfc.isis.ibex.ui.motor/build.properties
@@ -3,5 +3,4 @@ output.. = bin/
 bin.includes = META-INF/,\
                .,\
                plugin.xml,\
-               icons/,\
-               opis/
+               icons/
