+++ base/uk.ac.stfc.isis.ibex.ui.motor/plugin.xml
@@ -33,138 +33,10 @@
       </view>
    </extension>
    <extension
-         point="org.eclipse.ui.perspectives">
-      <perspective
-            class="uk.ac.stfc.isis.ibex.ui.motor.Perspective"
-            id="uk.ac.stfc.isis.ibex.ui.motor.perspective"
-            name="Main motors">
-      </perspective>
-   </extension>
-   <extension
-         point="org.eclipse.ui.perspectives">
-      <perspective
-            class="uk.ac.stfc.isis.ibex.ui.motor.Perspective"
-            id="uk.ac.stfc.isis.ibex.ui.motor.perspective"
-            name="Additional motors">
-      </perspective>
-   </extension>
-   <extension
-         point="org.eclipse.ui.perspectiveExtensions">
-      <perspectiveExtension
-            targetID="uk.ac.stfc.isis.ibex.ui.motor.perspective">
-         <view
-               closeable="false"
-               id="uk.ac.stfc.isis.ibex.ui.motor.views.MainMotorsView"
-               minimized="false"
-               ratio="0.1f"
-               relationship="right"
-               relative="uk.ac.stfc.isis.ibex.ui.perspectives.PerspectiveSwitcher"
-               showTitle="true"
-               standalone="true"
-               visible="true">
-         </view>
-         <view
-               closeable="false"
-               id="uk.ac.stfc.isis.ibex.ui.motor.views.AdditionalMotorsView"
-               minimized="false"
-               ratio="0.1f"
-               relationship="stack"
-               relative="uk.ac.stfc.isis.ibex.ui.motor.views.MainMotorsView"
-               showTitle="true"
-               standalone="true"
-               visible="true">
-         </view>
-         <view
-               closeable="false"
-               id="uk.ac.stfc.isis.ibex.ui.motor.views.AdditionalMotorsView2"
-               minimized="false"
-               ratio="0.1f"
-               relationship="stack"
-               relative="uk.ac.stfc.isis.ibex.ui.motor.views.AdditionalMotorsView"
-               showTitle="true"
-               standalone="true"
-               visible="true">
-         </view>
-         <view
-               closeable="true"
-               id="uk.ac.stfc.isis.ibex.ui.motor.views.MotorOPIView:*"
-               minimized="false"
-               moveable="false"
-               ratio="0.60f"
-               relationship="right"
-               relative="uk.ac.stfc.isis.ibex.ui.motor.views.MainMotorsView"
-               showTitle="true"
-               standalone="true"
-               visible="false">
-         </view>
-      </perspectiveExtension>
-   </extension>
-   <extension
          point="org.eclipse.ui.commands">
-      <command
-            defaultHandler="uk.ac.stfc.isis.ibex.ui.motor.displayoptions.DisplaySettingsHandler"
-            id="uk.ac.stfc.isis.ibex.ui.motor.colourOptionsCommand"
-            name="ColourOptions">
-         <commandParameter
-               id="org.eclipse.ui.commands.radioStateParameter"
-               name="colourOptions"
-               optional="false">
-         </commandParameter>
-         <state
-               id="org.eclipse.ui.commands.radioState">
-            <class
-                  class="org.eclipse.ui.handlers.RadioState">
-               <parameter
-                     name="default"
-                     value="NORMAL_VISION">
-               </parameter>
-               <parameter
-                     name="persisted"
-                     value="false">
-               </parameter>
-            </class>
-         </state>
-      </command>
    </extension>
    <extension
          point="org.eclipse.ui.menus">
-      <menuContribution
-            allPopups="false"
-            locationURI="menu:org.eclipse.ui.main.menu?after=runcontrolMenu">
-         <menu
-               id="preferencesMenu"
-               label="Preferences"
-               mnemonic="P">
-            <menu
-                  id="colourSettingsMenu"
-                  label="Colour Settings"
-                  mnemonic="C">
-               <command
-                     commandId="uk.ac.stfc.isis.ibex.ui.motor.colourOptionsCommand"
-                     id="uk.ac.stfc.isis.ibex.ui.motor.colourOptionsCommand1"
-                     label="Standard Colour Scheme"
-                     mnemonic="S"
-                     style="radio">
-                  <parameter
-                   		name="org.eclipse.ui.commands.radioStateParameter"
-                        value="NORMAL_VISION">
-                  </parameter>
-                       
-               </command>
-               <command
-                     commandId="uk.ac.stfc.isis.ibex.ui.motor.colourOptionsCommand"
-                     id="uk.ac.stfc.isis.ibex.ui.motor.colourOptionsCommand2"
-                     label="Alternative Colour Scheme 1"
-                     mnemonic="A"
-                     style="radio">
-                  <parameter
-                        name="org.eclipse.ui.commands.radioStateParameter"
-                        value="COLOURBLIND">
-                  </parameter>
-               </command>
-            </menu>
-         </menu>
-      </menuContribution>
    </extension>
 
 </plugin>
