+++ base/uk.ac.stfc.isis.ibex.ui/src/uk/ac/stfc/isis/ibex/ui/ImageUtils.java
@@ -7,14 +7,19 @@ import org.eclipse.swt.widgets.Display;
 /**
  * Utilities for SWT images.
  */
-public class ImageUtils {
-	/**
-	 * Gives an image a "disabled" look by greying it out.
-	 * 
-	 * @param source the source image
-	 * @return the image with a disabled look
-	 */
+public final class ImageUtils {
+    
+    private ImageUtils() {
+	
+    }
+    
+    /**
+     * Gives an image a "disabled" look by greying it out.
+     * 
+     * @param source the source image
+     * @return the image with a disabled look
+     */
     public static Image disabled(final Image source) {
-    	return new Image(Display.getDefault(), source, SWT.IMAGE_DISABLE);
+	return new Image(Display.getDefault(), source, SWT.IMAGE_DISABLE);
     }
 }
