+++ base/uk.ac.stfc.isis.scriptgenerator.product/META-INF/MANIFEST.MF
@@ -10,3 +10,4 @@ Require-Bundle: org.eclipse.ui;bundle-version="3.111.0",
  org.eclipse.core.runtime;bundle-version="3.15.100",
  org.eclipse.core.resources;bundle-version="3.13.200",
  uk.ac.stfc.isis.ibex.logger;bundle-version="1.0.0"
+Export-Package: uk.ac.stfc.isis.ibex.scriptgenerator.product
