+++ base/uk.ac.stfc.isis.scriptgenerator.product/build.properties
@@ -1,7 +1,5 @@
 bin.includes = plugin.xml,\
                META-INF/,\
                .,\
-               config.ini,\
-               icons/,\
-               splash.bmp
+               icons/
 source.. = src/
