+++ base/uk.ac.stfc.isis.scriptgenerator.product/src/uk/ac/stfc/isis/ibex/scriptgenerator/product/ApplicationWorkbenchAdvisor.java
@@ -23,6 +23,12 @@ import org.eclipse.ui.application.IWorkbenchWindowConfigurer;
 import org.eclipse.ui.application.WorkbenchAdvisor;
 import org.eclipse.ui.application.WorkbenchWindowAdvisor;
 
+/**
+ * Workbench advisor for Script generator application.
+ * 
+ * Contains hooks for application lifecycle.
+ *
+ */
 public class ApplicationWorkbenchAdvisor extends WorkbenchAdvisor {
 
     org.eclipse.ui.application.IWorkbenchConfigurer configurer;
